test: split compiled firewall base matching tests
